# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Ingredient {
  name: string;
  quantity: string;
  aisle: string;
  shoppingUrl: string;
  /** @minimum 0 */
  price: number;
}

export interface Recipe {
  id: number;
  title: string;
  description: string;
  imageUrl: string;
  category: string;
  dietary: string[];
  prepMinutes: number;
  cookMinutes: number;
  servings: number;
  calories: number;
  protein: number;
  carbs: number;
  fat: number;
  /** @minimum 0 */
  mealPrice: number;
  /** @minimum 0 */
  wholeItemCost: number;
  ingredients: Ingredient[];
  instructions: string[];
  author: string;
  createdAt: string;
}

export interface RecipeInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  imageUrl: string;
  /** @minLength 1 */
  category: string;
  dietary: string[];
  /** @minimum 0 */
  prepMinutes: number;
  /** @minimum 0 */
  cookMinutes: number;
  /** @minimum 1 */
  servings: number;
  /** @minimum 0 */
  calories: number;
  /** @minimum 0 */
  protein: number;
  /** @minimum 0 */
  carbs: number;
  /** @minimum 0 */
  fat: number;
  ingredients: Ingredient[];
  instructions: string[];
  /** @minLength 1 */
  author: string;
}

export interface RecipeUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  imageUrl?: string;
  category?: string;
  dietary?: string[];
  prepMinutes?: number;
  cookMinutes?: number;
  servings?: number;
  calories?: number;
  protein?: number;
  carbs?: number;
  fat?: number;
  ingredients?: Ingredient[];
  instructions?: string[];
  author?: string;
}

export interface RecipeSummary {
  totalRecipes: number;
  totalCategories: number;
  averageCalories: number;
  popularCategory: string;
  totalIngredients: number;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type VideoRecipeImportRequestSourceType = typeof VideoRecipeImportRequestSourceType[keyof typeof VideoRecipeImportRequestSourceType];


export const VideoRecipeImportRequestSourceType = {
  url: 'url',
  upload: 'upload',
} as const;

export interface VideoRecipeImportRequest {
  sourceType: VideoRecipeImportRequestSourceType;
  /** @minLength 1 */
  source: string;
  mimeType?: string;
}

export interface VideoRecipeImportResponse {
  recipe: RecipeInput;
  sourceLabel: string;
}

export type VideoRecipeImportJobResponseStatus = typeof VideoRecipeImportJobResponseStatus[keyof typeof VideoRecipeImportJobResponseStatus];


export const VideoRecipeImportJobResponseStatus = {
  processing: 'processing',
  complete: 'complete',
  failed: 'failed',
} as const;

export interface VideoRecipeImportJobResponse {
  jobId: string;
  status: VideoRecipeImportJobResponseStatus;
  recipe?: RecipeInput;
  sourceLabel?: string;
  error?: string;
}

export type ListRecipesParams = {
search?: string;
category?: string;
dietary?: string;
/**
 * @minimum 1
 */
maxCalories?: number;
};