# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /recipes:
    get:
      operationId: listRecipes
      tags: [recipes]
      summary: List recipes
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: dietary
          in: query
          required: false
          schema:
            type: string
        - name: maxCalories
          in: query
          required: false
          schema:
            type: number
            minimum: 1
      responses:
        "200":
          description: Recipes matching the filters
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipe"
    post:
      operationId: createRecipe
      tags: [recipes]
      summary: Add a recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeInput"
      responses:
        "201":
          description: Recipe created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "400":
          description: Invalid recipe
  /recipes/{id}:
    get:
      operationId: getRecipe
      tags: [recipes]
      summary: Get a recipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Recipe details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "404":
          description: Recipe not found
    patch:
      operationId: updateRecipe
      tags: [recipes]
      summary: Update a recipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipeUpdate"
      responses:
        "200":
          description: Recipe updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        "404":
          description: Recipe not found
    delete:
      operationId: deleteRecipe
      tags: [recipes]
      summary: Delete a recipe
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Recipe deleted
        "404":
          description: Recipe not found
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [Storage]
      summary: Request a presigned URL for a cooking video upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Invalid upload metadata
        "401":
          description: Authentication required
  /storage/objects/{objectPath}:
    get:
      operationId: getStorageObject
      tags: [Storage]
      summary: Serve an uploaded object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found
components:
  schemas:
...
    Ingredient:
      type: object
      properties:
        name:
          type: string
        quantity:
          type: string
        aisle:
          type: string
        shoppingUrl:
          type: string
        price:
          type: number
          minimum: 0
      required: [name, quantity, aisle, shoppingUrl, price]
    Recipe:
      type: object
      properties:
        id:
          type: number
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        category:
          type: string
        dietary:
          type: array
          items:
            type: string
        prepMinutes:
          type: number
        cookMinutes:
          type: number
        servings:
          type: number
        calories:
          type: number
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
        mealPrice:
          type: number
          minimum: 0
        wholeItemCost:
          type: number
          minimum: 0
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/Ingredient"
        instructions:
          type: array
          items:
            type: string
        author:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - imageUrl
        - category
        - dietary
        - prepMinutes
        - cookMinutes
        - servings
        - calories
        - protein
        - carbs
        - fat
        - mealPrice
        - wholeItemCost
        - ingredients
        - instructions
        - author
        - createdAt
    RecipeInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        imageUrl:
          type: string
        category:
          type: string
          minLength: 1
        dietary:
          type: array
          items:
            type: string
        prepMinutes:
          type: number
          minimum: 0
        cookMinutes:
          type: number
          minimum: 0
        servings:
          type: number
          minimum: 1
        calories:
          type: number
          minimum: 0
        protein:
          type: number
          minimum: 0
        carbs:
          type: number
          minimum: 0
        fat:
          type: number
          minimum: 0
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/Ingredient"
        instructions:
          type: array
          items:
            type: string
        author:
          type: string
          minLength: 1
      required:
        - title
        - description
        - imageUrl
        - category
        - dietary
        - prepMinutes
        - cookMinutes
        - servings
        - calories
        - protein
        - carbs
        - fat
        - ingredients
        - instructions
        - author
    RecipeUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        imageUrl:
          type: string
        category:
          type: string
        dietary:
          type: array
          items:
            type: string
        prepMinutes:
          type: number
        cookMinutes:
          type: number
        servings:
          type: number
        calories:
          type: number
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
        ingredients:
          type: array
          items:
            $ref: "#/components/schemas/Ingredient"
        instructions:
          type: array
          items:
            type: string
        author:
          type: string
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: number
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"
...