# lib/api-zod/src/generated/types/ingredient.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Ingredient {
  name: string;
  quantity: string;
  aisle: string;
  shoppingUrl: string;
  /** @minimum 0 */
  price: number;
}