# lib/api-zod/src/generated/types/recipeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Ingredient } from './ingredient';

export interface RecipeInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  imageUrl: string;
  /** @minLength 1 */
  category: string;
  dietary: string[];
  /** @minimum 0 */
  prepMinutes: number;
  /** @minimum 0 */
  cookMinutes: number;
  /** @minimum 1 */
  servings: number;
  /** @minimum 0 */
  calories: number;
  /** @minimum 0 */
  protein: number;
  /** @minimum 0 */
  carbs: number;
  /** @minimum 0 */
  fat: number;
  ingredients: Ingredient[];
  instructions: string[];
  /** @minLength 1 */
  author: string;
}