# lib/api-zod/src/generated/types/recipeUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Ingredient } from './ingredient';

export interface RecipeUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  imageUrl?: string;
  category?: string;
  dietary?: string[];
  prepMinutes?: number;
  cookMinutes?: number;
  servings?: number;
  calories?: number;
  protein?: number;
  carbs?: number;
  fat?: number;
  ingredients?: Ingredient[];
  instructions?: string[];
  author?: string;
}